Update additional_config.conf
Removed the golang binary that was used to poll interface data.  I still think there's value in getting the data from the network stack rather than pfSense, but it's working for now.  Maybe I'll add it as an additional influx interface source after I mess with it a little more.