Cleaning up Grafana Cruft
Grafana keeps raw queries in the json and it was bugging me.  If the raw query isn't displayed, why save it?  Why not strip it on export for external use?  Also, my interfaces were embedded in the json.  No security risk, but it bugged me.  This is supposed to be templated for external use!